refactor: 패키지 변경 (car, engine, race, utils)
