+++ src/main/java/racing/car/CarNames.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.car;
 
 import java.util.ArrayList;
 import java.util.List;
