+++ src/main/java/racing/car/CarStatus.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.car;
 
 public enum CarStatus {
 	FORWARD, STOP;
