+++ src/main/java/racing/car/engine/Power.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.car.engine;
 
 public class Power {
 	private int val;
