+++ src/test/java/racing/utils/StringUtilsTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.utils;
 
 import static org.junit.jupiter.api.Assertions.*;
